pyproject.toml
[project]
name = "aula20-bootcamp-python"
version = "0.1.0"
description = ""
authors = [
    {name = "Renato"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "streamlit (>=1.42.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
